langchain/review_agent/requirements.txt
chromadb
langchain-huggingface
sentence_transformers
langchain
python-dotenv
pandas
langchain-chroma
langchain_groq